Split session test utils & some fixes
