+++ crates/session/Cargo.toml
@@ -23,7 +23,6 @@ uuid = { version = "1", features = ["serde", "v4"] }
 [dev-dependencies]
 anyhow = "1"
 serial_test = "3"
-smplx-regtest = "0.0.5"
 sqlx = { version = "0.8.3", features = ["postgres", "runtime-tokio-rustls", "migrate"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 wiremock = "0.6"
